Feat: Adds metrics server port configuration (#88)
* Feat: Adds metrics server port configuration

Allows configuration of the metrics server port via a command-line argument.
The default port is 9001 if no port is specified.

Removes unused metrics and simplifies logging.

* Fix: Adds ignore attribute to a test

Adds an ignore attribute to the test to prevent it from running in CI without a live node.

Removes the target cache mount from one of the docker build steps.

* Style💎: Cargofmt

Removes unnecessary variables in Engine creation for conciseness.

Removes an extraneous newline in the store and metrics files.

* fix: cargo fmt errors and unresolved module imports